requirements.txt
# Week 1 라이트봇 A단계 - 필수 패키지
# 데이터베이스
psycopg2-binary==2.9.7

# 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 비동기 처리
celery==5.3.4
redis==5.0.1

# HTTP 클라이언트
httpx==0.25.2
aiohttp==3.9.1

# 데이터 처리
pandas==2.1.3
numpy==1.25.2

# 환경변수
python-dotenv==1.0.0

# 로깅
structlog==23.2.0

# 유틸리티
pydantic==2.5.0
python-dateutil==2.8.2

# 개발 도구 (선택사항)
pytest==7.4.3
pytest-asyncio==0.21.1
